ng-user [ref] renamed /user/user-tile to /user/tile - fix some shjenanigans
