+++ packages/ng/src/app/user/index.ts
@@ -1,3 +1,5 @@
-export * from './user-tile';
-export { LuUserModule } from './user.module';
 export { IUser } from './user.model';
+export * from './tile';
+export * from './display';
+export * from './picture';
+export { LuUserModule } from './user.module';
