+++ packages/ng/src/app/user/tile/user-tile.component.scss
