+++ packages/ng/src/app/user/user.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { LuUserTileModule } from './user-tile';
+import { LuUserTileModule } from './tile';
 import { LuUserPictureModule } from './picture';
 import { LuUserDisplayModule } from './display';
 
@@ -16,6 +16,7 @@ import { LuUserDisplayModule } from './display';
 	exports: [
 		LuUserPictureModule,
 		LuUserDisplayModule,
+		LuUserTileModule,
 	]
 })
 export class LuUserModule { }
